Move arraysHaveSameElements function to its own file
